Move again data files
