+++ src/moscot/utils/__init__.py
@@ -1,1 +1,1 @@
-from . import subset_policy, tagged_array
+from . import data, subset_policy, tagged_array
