+++ src/moscot/utils/_data/allTFs_mm.txt
