+++ src/moscot/utils/_data/mouse_apoptosis.txt
