+++ src/moscot/utils/_data/mouse_proliferation.txt
